refactor(kodando-redurx): Renamed to `kodando-store` [breaking].
