+++ kodando-store/build.gradle
@@ -1,4 +1,4 @@
-version '0.7.0'
+version '0.1.0'
 
 dependencies {
     compile project(":kodando-rxjs")
