+++ kodando-store/src/main/kotlin/kodando/store/Action.kt
@@ -1,4 +1,4 @@
-package kodando.redurx
+package kodando.store
 
 interface Action {
   @JsName("type")
