+++ kodando-store/src/main/kotlin/kodando/store/BaseAction.kt
@@ -1,4 +1,4 @@
-package kodando.redurx
+package kodando.store
 
 abstract class BaseAction : Action {
   override val type = this::class.simpleName!!
