+++ kodando-store/src/main/kotlin/kodando/store/GlobalActionDispatcher.kt
@@ -1,4 +1,4 @@
-package kodando.redurx
+package kodando.store
 
 import kodando.rxjs.Observable
 import kodando.rxjs.Subject
