+++ kodando-store/src/main/kotlin/kodando/store/Reducer.kt
@@ -1,3 +1,3 @@
-package kodando.redurx
+package kodando.store
 
 typealias Reducer<T> = (T, Action) -> T
