+++ kodando-store/src/test/kotlin/kodando/store/EffectSpec.kt
@@ -1,4 +1,4 @@
-package kodando.redurx
+package kodando.store
 
 import kodando.jest.Spec
 import kodando.jest.expect
