+++ kodando-store/src/test/kotlin/kodando/store/Main.kt
@@ -1,4 +1,4 @@
-package kodando.redurx
+package kodando.store
 
 fun main(args: Array<String>) {
   EffectSpec()
