+++ settings.gradle
@@ -10,7 +10,7 @@ include 'kodando-react-native'
 include 'kodando-react-router-dom'
 include 'kodando-elmish'
 include 'kodando-elmish-react'
-include 'kodando-redurx'
+include 'kodando-store'
 include 'kodando-mithril'
 include 'kodando-mithril-router-dom'
 include 'kodando-mobx'
